requirements.txt
chromadb
groq
python-dotenv
PyPDF2
python-docx
pypdf
